fix : fix sub errors
