+++ app/userDetail/_layout.tsx
@@ -1,20 +1,5 @@
-import { COLORS, FONT_SIZE, FONTS } from "@/shared/styles";
-import { FilterDefaultProvider } from "@/widgets/home/contexts";
 import { Stack } from "expo-router";
 
 export default function UserDetailLayout() {
-  return (
-    <FilterDefaultProvider>
-      <Stack
-        screenOptions={{
-          headerShown: false,
-          headerTitleStyle: { fontFamily: FONTS.bold, fontSize: FONT_SIZE.b1 },
-          contentStyle: { backgroundColor: COLORS.white },
-        }}
-      >
-        <Stack.Screen name="[userId]" options={{ headerShown: true, headerTitle: "유저 프로필" }}/>
-
-      </Stack>
-    </FilterDefaultProvider>
-  );
+  return <Stack screenOptions={{ headerShown: false }} />;
 }
