+++ shared/api/request.ts
@@ -10,6 +10,7 @@ export const REQUEST = {
   USER_LIKED: "/likes",
   USER_ALL: "/profile/all",
   USER_SEARCH: "/profile/filter",
+  USER_RECOMMEND: "/profile/recommend",
   DASHBOARD_NOROOM: "/dashboard/noroom",
   DASHBOARD_ROOM: "/dashboard/hasroom",
 };
