+++ widgets/home/components/index.ts
@@ -9,3 +9,18 @@ export { default as TabHeader } from "./TabHeader";
 export { default as UserFilterBottomSheet } from "./UserFilterBottomSheet";
 export { default as UserListItem } from "./UserListItem";
 export { default as UserSearchFilter } from "./UserSearchFilter";
+
+// Home screen components
+export { ActionCard } from "./ActionCard";
+export {
+  EmptyFavoriteUsersState,
+  EmptyRecommendUsersState,
+} from "./EmptyStates";
+export { FavoriteUsersSection } from "./FavoriteUsersSection";
+export { Header } from "./Header";
+export { RecommendSection } from "./RecommendSection";
+export { RecommendUserItem } from "./RecommendUserItem";
+export { SurveyNotification } from "./SurveyNotification";
+export { UserDetailModal } from "./UserDetailModal";
+export { RecommendSkeleton, UserProfileCard } from "./UserProfileCard";
+export { WelcomeSection } from "./WelcomeSection";
